requirements.txt
ansicon==1.89.0
blessed==1.19.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.7
future==0.18.2
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
inquirer==2.8.0
jinxed==1.1.0
oauthlib==3.1.0
packaging==20.9
protobuf==3.15.6
pwinput==1.0.2
pyad==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pypiwin32==223
python-editor==1.0.4
python-redmine==2.3.0
pytz==2021.1
pywin32==302
readchar==2.0.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
stdiomask==0.0.6
tpm==3.6
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
